+++ src/runway/commands/runway/__init__.py
@@ -6,7 +6,6 @@ from .gen_sample import *  # noqa
 from .gitclean import *  # noqa
 from .init import *  # noqa
 from .preflight import *  # noqa
-from .run_python import *  # noqa
 from .run_stacker import *  # noqa
 from .test import *  # noqa
 from .whichenv import *  # noqa