+++ src/runway/tests/handlers/__init__.py
@@ -1,0 +1,7 @@
+"""Import classes."""
+
+# pylint: disable = wildcard-import
+
+from .cfn_lint import *  # noqa
+from .script import *  # noqa
+from .yaml_lint import *  # noqa